feat: Divided Google Map Components
